fetch data in about page

+++ admin/view_testimonial.php
@@ -78,7 +78,7 @@
                                             <a href="details_testimonial.php?id=<?php echo $row->test_id; ?>" class="btn btn-success btn-sm">View</a>
                                         </td>
                                         <td>
-                                            <a href="delete_testimonial.php?id=<?php echo $row->test_id; ?>" class="btn btn-danger btn-sm">Delete</a>
+                                            <a onclick="javascript:return confirm('Are You Sure?')" href="delete_testimonial.php?id=<?php echo $row->test_id; ?>" class="btn btn-danger btn-sm">Delete</a>
                                         </td>
                                     </tr>
                                 <?php
